+++ ChangeLog.md
@@ -1,7 +1,8 @@
 2020-04
 
 - preliminaries for an ecoz2py module based on cffi
-    - exposing hmm_learn but having some weird cffi issues
+- expose hmm_learn
+    
 - in general module to be imported as `import ecoz2py as ecoz2`
   and functions called as eg. `ecoz2.prd_show_file(...)`